% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdep.R
\name{gcor-package}
\alias{gcor-package}
\alias{gcor}
\alias{dgcor}
\alias{gdis}
\title{Estimate generalized correlation and related measures}
\usage{
gcor(x, y = NULL, k = NULL, data = NULL, simplify = TRUE, dropNA = "none")

dgcor(x, y = NULL, k = NULL, data = NULL, simplify = TRUE, dropNA = "none")

gdis(x, k = NULL, dropNA = "none", ...)
}
\arguments{
\item{x}{a vector, matrix, data frame or formula. If formula, \code{data} should be specified.
\code{gdis} requires a matrix or data frame.}

\item{y}{\code{NULL} (default) or a vector, matrix or data frame with compatible dimensions to \code{x}.}

\item{k}{\code{NULL} (default) or an integer specifying the number of groups for discretization.
Numerical data are divided into \code{k} groups using \code{k}-quantiles.
If \code{NULL}, it is determined automatically.}

\item{data}{\code{NULL} (default) or a data frame. Required if \code{x} is a formula.}

\item{simplify}{a logical. If \code{TRUE}, the returned value is coerced to
a vector when one of its dimensions is one.}

\item{dropNA}{a character specifying how to handle missing values.
It should be one of the following:
\describe{
\item{\code{"none"}}{(default) Treat missing values as observations of a single
categorical value, namely \code{NA}. Recommended for reflecting missing patterns in the analysis.}
\item{\code{"casewise"}}{Casewise deletion; rows containing any missing value are removed.
Similar to \code{use = "complete.obs"} in \code{\link{cor}}.}
\item{\code{"pairwise"}}{Pairwise deletion; for any pair of columns (x,y),
the i-th row is removed if x[i] or y[i] is missing.
This process is applied per pair; the same row is not removed with another pair (w,z),
if both w[i] and z[i] are not missing.
Similar to \code{use = "pairwise.complete.obs"} in \code{\link{cor}}.}
}}

\item{...}{additional arguments (\code{diag} and \code{upper}) passed to \code{as.dist} function.
See \code{\link{as.dist}} for details.}
}
\value{
For \code{gcor} and \code{dgcor}, a numeric matrix is returned (or a vector if \code{simplify = TRUE}).
For \code{gdis}, an object of class \code{"dist"} is returned.
}
\description{
Estimate measures based on mutual dependency, which includes:
\itemize{
\item{Generalized correlation measure (\code{gcor})}
\item{Directed generalized correlation measure (\code{dgcor})}
\item{Dissimilarity between variables (\code{gdis})}
}
}
\examples{
# Generalized correlation measure
gcor(iris)

# Dependency of Species on other variables
dgc <- dgcor(Species ~ ., data = iris)
dotchart(sort(dgc), main = "Predictability of Species")

# Clustering
gd <- gdis(iris)
hc <- hclust(gd, method = "ward.D2")
plot(hc)

# Multidimensional scaling
mds <- cmdscale(gd, k = 2)
plot(mds, type = "n", xlab = "", ylab = "", asp = 1, axes = FALSE,
     main = "cmdscale with gdis(iris)")
text(mds[,1], mds[,2], rownames(mds))
}
\references{
Suzuki, R. (2025). \emph{A generalization of correlation coefficient}. preprint.
\url{https://r-suzuki.github.io}
}
